Add files via upload
Group Project: Covid-19 Data Analysis 

 

Members: 	Marco Zanella (16-621-419) 

Andrin Hofstetter (16-610-586) 

Ivan Theiler (16-619-652) 

 

Goal:  

The goal of our project is to facilitate the processing of data about the Covid-19 pandemic. For this purpose we wanted to develop a program that provides a selection of countries and possible information from which the user can create his own data set. In addition, there is the possibility to make a comparison with the development in Switzerland. 

 

Manual: 

The project folder contains this manual, the actual python program (code) and two CSV files. It is important to link the python program with the mentioned folder, so the code can resort the necessary data from the CSV file. After starting the code, the user will be asked to choose his desired country and the information the user wants to have deducted. We limited the countries as we wanted to show the user the possibilities/options, so it is still comprehensible. Secondly, the user can choose multiple key values he wants to receive from the program about the country. On-one-hand, Covid-19 specific information, as well as general interesting values like the GDP or the life expectacy. Next, the user can chose the desired date (with the current dataset between 24.02.2020 and 11.05.2021) he wants to receive the information from. Additionally, we added the possibility to compare it directly with Switzerland. The information are comprehensibly presented and provide a fast & efficient value-added. Simultaneously, the program inserts and stores all the requests made from the user in the second CSV file. This is an essential part, as the user is then able to use the data easily for further research & use.  

The Covid-19 data was derived from https://github.com/owid/covid-19-data/blob/master/public/data/README.md 

Future Adaptations: 

It would be an amazing extension of the program to add an automation of the data update, so the new data published in the web could be implemented in the dataset, so the user is always able to request new dates. Additionally, the dataset can always be enlarged with new options, new countries and other possibilities.  

 

Difficulties: 

The biggest challenge for us was to get a uniform database, as there are many different versions for each country, each option and key numbers. Accordingly, we have limited the selection only to countries that publish the data according to our requirements.  

 

Summary: 

Thanks to the programming course, we learned a lot of new techniques and realised the possibilities and power of coding. During the project we really blossomed and plan to continue and use the new skills in the future.